✨ [BE-S13P11A403-75] 인기글을 제외한 게시글 구현 완료

+++ backend/bbatty/src/main/java/com/ssafy/bbatty/domain/board/dto/response/PostDetailResponse.java
@@ -11,8 +11,8 @@ public class PostDetailResponse {
     private String title;
     private String authorNickname;
     private String content;
-    private Long likeCount;
-    private Long viewCount;
+    private Integer likeCount;
+    private Integer viewCount;
     private Integer commentCount;
     private String createdAt;
     private String updatedAt;
